requirements.txt
amqp==2.6.1
asgiref==3.2.10
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==4.4.7
Django==3.1
django-appconf==1.0.4
django-bootstrap4==2.2.0
django-celery-email==3.0.0
django-crispy-bootstrap==0.1.1.1
importlib-metadata==1.7.0
jedi==0.17.2
kombu==4.6.11
parso==0.7.1
pluggy==0.13.1
python-jsonrpc-server==0.3.4
python-language-server==0.34.1
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
ujson==1.35
vine==1.3.0
whitenoise==5.2.0
zipp==3.1.0
psycopg2-binary==2.8.5
django-bootstrap-datepicker-plus==3.0.5
holidays==0.10.3

